requirements.txt
fastapi
uvicorn
requests
python-docx
pymupdf
python-dotenv
pydantic
whoosh
openpyxl
xlrd
python-pptx
